+++ sqlx-exasol-impl/src/connection/etl/mod.rs
@@ -38,7 +38,9 @@
 //! let mut con = pool.acquire().await?;
 //! let mut tx = con.begin().await?;
 //!
-//! let (query_fut, writers) = ImportBuilder::new("SOME_TABLE").build(&mut *tx).await?;
+//! let (query_fut, writers) = ImportBuilder::new("SOME_TABLE", None)
+//!     .build(&mut *tx)
+//!     .await?;
 //!
 //! // concurrently use the writers and await the query future
 //!